requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
openai
llama-index
python-dotenv
langchain_groq
langchain-deepseek
llama-index-embeddings-google-genai
pydantic
llama-index-vector-stores
tiktoken
httpx
requests
rich
langchain-google-genai
tavily
sqlite